+++ src/ansys/mechanical/core/__init__.py
@@ -2,6 +2,7 @@
 import logging
 import os
 
+from ansys.tools.path import find_mechanical
 import appdirs
 
 # Setup data directory
@@ -23,7 +24,6 @@ from ansys.mechanical.core._version import __version__
 from ansys.mechanical.core.mechanical import (
     change_default_mechanical_path,
     close_all_local_instances,
-    find_mechanical,
     get_mechanical_path,
     launch_mechanical,
 )