+++ tests/test_mechanical.py
@@ -5,6 +5,7 @@ import re
 import subprocess
 import sys
 
+import ansys.tools.path
 import grpc
 import pytest
 
@@ -414,7 +415,7 @@ def test_launch_result_mode(mechanical_result):
 @pytest.mark.remote_session_launch
 def test_find_mechanical_path():
     if pymechanical.mechanical.get_start_instance():
-        path, version = pymechanical.find_mechanical()
+        path, version = ansys.tools.path.find_mechanical()
 
         if misc.is_windows():
             assert "AnsysWBU.exe" in path
@@ -427,11 +428,11 @@ def test_find_mechanical_path():
 @pytest.mark.remote_session_launch
 def test_change_default_mechanical_path():
     if pymechanical.mechanical.get_start_instance():
-        path, version = pymechanical.find_mechanical()
+        path, version = ansys.tools.path.find_mechanical()
 
         pymechanical.change_default_mechanical_path(path)
 
-        path_new, version_new = pymechanical.find_mechanical()
+        path_new, version_new = ansys.tools.path.find_mechanical()
 
         assert path_new == path
         assert version_new == version
@@ -440,17 +441,17 @@ def test_change_default_mechanical_path():
 @pytest.mark.remote_session_launch
 def test_version_from_path():
     windows_path = "C:\\Program Files\\ANSYS Inc\\v231\\aisol\\bin\\winx64\\AnsysWBU.exe"
-    version = pymechanical.mechanical._version_from_path(windows_path)
+    version = ansys.tools.path.version_from_path("mechanical", windows_path)
     assert version == 231
 
     linux_path = "/usr/ansys_inc/v231/aisol/.workbench"
-    version = pymechanical.mechanical._version_from_path(linux_path)
+    version = ansys.tools.path.version_from_path("mechanical", linux_path)
     assert version == 231
 
     with pytest.raises(RuntimeError):
         # doesn't contain version
         path = "C:\\Program Files\\ANSYS Inc\\aisol\\bin\\winx64\\AnsysWBU.exe"
-        pymechanical.mechanical._version_from_path(path)
+        ansys.tools.path.version_from_path("mechanical", path)
 
 
 @pytest.mark.remote_session_launch